interactive: rework setting of tx and rx frequencies
The previous implementation had a bug where unless you entered 7
characters, including newlines, it would forever loop. Instead we now
set a timeout of 10 seconds, and use Serial.parseInt which reads an
integer from the serial port until a non-numeric character is found.

While we're here, split out common functionality into helper functions.
This saves us a few hundred bytes of flash. It remains to be seen if
this is a bad trade off for stack consumption at runtime. Execution time
may suffer slightly, but it is not important here.

Signed-off-by: Joel Stanley <joel@jms.id.au>