interactive: Add channel selection
Using the new menu item 'A' (case insensitive), a user can select from a
list of predefined channels. This eases configuration when setting up
more than one device, with the TX being 1500Hz above the RX.

Note that we listen on the third harmonic of the TX to avoid issues with
the RX stage not be able to tune very well above 27MHz (due to phase
imbalance issues).

Signed-off-by: Joel Stanley <joel@jms.id.au>