interactive: Workaround for old versions of GCC
The Arduino distribution for OSX ships an ancient version of GCC (4.3.2
vs 4.8.1 that Ubuntu has). This creates a binary that is 32% larger, and
appears to fail to access data placed in the PROGDATA section correctly.

Add a workaround so users do not hit this bug. When we get to the bottom
of it, we can remove the SRAM version and leave the lookup table in
flash.

Signed-off-by: Joel Stanley <joel@jms.id.au>